Initial commit
First submission of the first capstone project. 
NOTE: there is a HUGE comment block in the middle of the text asking the grader a question. I will definitely need to remove that and resubmit, but I did not know how to explain my concern/ask my question. Obviously a comment block THAT BIG is not appropriate in a program like this.