Initial commit
Three programs, two required, one optional challenge. The John.py program stores a growing list of wrong names and stops when the name 'John' is printed. cafe.py has information of the stock amount and prices and computes the total value in the store using dictionaries. The challenge question requires the usage of the 'enumerate' function.